requirements.txt
streamlit
streamlit_extras
streamlit_aggrid
streamlit_antd_components
xinference-client
ollama
langchain
langchain_core
langchain-openai
langchain-milvus
langchain-chroma
langchain_community
langchain_text_splitters
langgraph
sqlalchemy
rapidocr-onnxruntime
python-docx
python-pptx
pymupdf
